client/requirements.txt
pyinstaller==6.10.0
pyautogui==0.9.54
Pillow==10.4.0
requests==2.32.3
keyboard==0.13.5